requirements.txt
asyncpg>=0.27.0
python-telegram-bot>=20.0
tzdata